Files relevant to database project.
The final project for databases was broken in multiples steps. First the development of an ER diagram to development of a script and tables. Because of the amount of data needed the script was quite long. All data was collected from either the CT DOT's website or from the US DOT's website.